+++ kwcoco/util/util_delayed_poc.py
@@ -3,18 +3,18 @@ This module has been reorganized into kwcoco.util.delayed_poc and will likely
 be further refactored.
 """
 
-from kwcoco.util.delayed_poc.delayed_base import (DelayedImageOperation,
-                                                  DelayedVideoOperation,
-                                                  DelayedVisionOperation,)
+from kwcoco.util.delayed_poc.delayed_base import (DelayedImage,
+                                                  DelayedVideo,
+                                                  DelayedVisionMixin,)
 from kwcoco.util.delayed_poc.delayed_leafs import (DelayedLoad, DelayedNans,
                                                    DelayedIdentity, dequantize)
-from kwcoco.util.delayed_poc.delayed_nodes import (DelayedChannelConcat,
+from kwcoco.util.delayed_poc.delayed_nodes import (DelayedChannelStack,
                                                    DelayedCrop,
-                                                   DelayedFrameConcat,
+                                                   DelayedFrameStack,
                                                    DelayedWarp, JaggedArray,)
 
 
-__all__ = ['DelayedChannelConcat', 'DelayedCrop', 'DelayedFrameConcat',
-           'DelayedIdentity', 'DelayedImageOperation', 'DelayedLoad',
-           'DelayedNans', 'DelayedVideoOperation', 'DelayedVisionOperation',
+__all__ = ['DelayedChannelStack', 'DelayedCrop', 'DelayedFrameStack',
+           'DelayedIdentity', 'DelayedImage', 'DelayedLoad',
+           'DelayedNans', 'DelayedVideo', 'DelayedVisionMixin',
            'DelayedWarp', 'JaggedArray', 'dequantize']